+++ coxeter/families/plane_shape_families.py
@@ -113,7 +113,9 @@ class TruncationPlaneShapeFamily(ShapeFamily):
         # those cases.
         dets = np.linalg.det(coeffs)
         solution_indices = np.abs(dets) > thresh
-        xs = np.linalg.solve(coeffs[solution_indices], bs[solution_indices])
+        xs = np.linalg.solve(coeffs[solution_indices], bs[solution_indices][..., None])[
+            ..., 0
+        ]
 
         # Reject any solutions that are intersections that lie beyond at least
         # one of the bounding planes.
@@ -441,7 +443,7 @@ class Family523(TruncationPlaneShapeFamily):
         """
         if not 1 <= a <= (cls.s * np.sqrt(5)):
             raise ValueError(
-                "The a parameter must be between 1 and s\u221A5 "
+                "The a parameter must be between 1 and s\u221a5 "
                 "(where s is the inverse of the golden ratio)."
             )
         if not cls.S**2 <= c <= 3: